requirements.txt
numpy
opencv-python
Pillow
tqdm
matplotlib
imgaug